# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creatrid API
  description: |
    Creatrid is a verified digital identity platform for creators. Connect social
    accounts, build a Creator Score, manage content licensing, and share a public
    profile.

    ## Authentication
    Most endpoints require authentication via an HTTP-only cookie named `token`
    that contains a JWT issued after Google OAuth sign-in.

    Third-party verification endpoints use API key authentication via the
    `X-API-Key` header.
  version: 1.0.0
  contact:
    name: Creatrid
    url: https://creatrid.com
servers:
  - url: https://creatrid-api.bluewave-351c9003.westeurope.azurecontainerapps.io
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/users/{username}/content:
    get:
      operationId: getPublicContent
      tags: [Users]
      summary: Get public content
      description: Returns the public content items for a creator by username.
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: type
          in: query
          schema:
            type: string
            enum: [image, video, audio, text, other]
          description: Filter by content type
        - name: q
          in: query
          schema:
            type: string
          description: Search query for title or description
      responses:
        "200":
          description: Public content list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContentItemPublic"
                  total:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/content:
    post:
      operationId: uploadContent
      tags: [Content]
      summary: Upload content
      description: Uploads a new content item to the vault. Supports images, videos, audio, PDFs, and other files up to 500 MB.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max 500 MB)
                title:
                  type: string
                  description: Content title
                description:
                  type: string
                  description: Content description
                tags:
                  type: string
                  description: Comma-separated tags
                is_public:
                  type: string
                  enum: ["true", "false"]
                  description: Whether the content is publicly visible (default true)
      responses:
        "201":
          description: Content uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "400":
          description: File too large or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: File upload not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listContent
      tags: [Content]
      summary: List my content
      description: Returns the authenticated user's content items (both public and private).
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Content list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContentItem"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/content/{id}:
    get:
      operationId: getContent
      tags: [Content]
      summary: Get content item
      description: Returns a single content item. Only the owner can access it.
      parameters:
        - $ref: "#/components/parameters/ContentID"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateContent
      tags: [Content]
      summary: Update content metadata
      description: Updates the title, description, tags, and visibility of a content item.
      parameters:
        - $ref: "#/components/parameters/ContentID"
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                tags:
                  type: array
                  items:
                    type: string
                isPublic:
                  type: boolean
      responses:
        "200":
          description: Content updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteContent
      tags: [Content]
      summary: Delete content
      description: Deletes a content item and its associated blob storage file.
      parameters:
        - $ref: "#/components/parameters/ContentID"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Content deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token set as an HTTP-only cookie after Google OAuth sign-in
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for third-party verification endpoints (starts with crk_)
  parameters:
    Username:
      name: username
      in: path
      required: true
      schema:
        type: string
      description: Creator username
...
    ContentID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Content item ID
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items to return (default 20, max 100)
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Not authenticated
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Not authorized
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Not found
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    ContentItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        contentType:
          type: string
          enum: [image, video, audio, text, other]
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
          description: File size in bytes
        fileUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        hashSha256:
          type: string
          description: SHA-256 hash of the original file
        isPublic:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContentItemPublic:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        contentType:
          type: string
          enum: [image, video, audio, text, other]
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        hashSha256:
          type: string
        isPublic:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...